[1520] Move course financial incentives logic to a concern
This makes it a bit clearer that this logic is separate from everything else.
