+++ app/models/course.rb
@@ -24,6 +24,7 @@
 
 class Course < ApplicationRecord
   include WithQualifications
+  include WithFinancialIncentives
   include ChangedAt
 
   has_associated_audits
@@ -271,31 +272,5 @@ class Course < ApplicationRecord
     to_remove.each { |site| remove_site!(site: site) }
 
     sites.reload
-  end
-
-  def has_bursary?
-    dfe_subjects.any?(&:has_bursary?)
-  end
-
-  def has_scholarship_and_bursary?
-    dfe_subjects.any?(&:has_scholarship_and_bursary?)
-  end
-
-  def has_early_career_payments?
-    dfe_subjects.any?(&:has_early_career_payments?)
-  end
-
-  def bursary_amount
-    dfe_subject_which_defines_financial_incentives.bursary_amount
-  end
-
-  def scholarship_amount
-    dfe_subject_which_defines_financial_incentives.scholarship_amount
-  end
-
-private
-
-  def dfe_subject_which_defines_financial_incentives
-    dfe_subjects.max_by(&:total_bursary_and_early_career_payments_amount)
   end
 end
